Extend interface governance.blockChain
